+++ CHANGELOG.md
@@ -9,6 +9,8 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+## [v1.17.0](https://github.com/allenai/beaker-py/releases/tag/v1.17.0) - 2023-02-13
+
 ### Added
 
 - Added `ExperimentSpec.new()` constructor.